add sql to response your rank

+++ database/migrations/2019_11_07_125522_create_rankings_table.php
@@ -16,7 +16,7 @@ class CreateRankingsTable extends Migration
         Schema::create('rankings', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('name');
-            $table->decimal('time', 8, 1);
+            $table->float('time', 8, 1);
             $table->timestamps();
         });
     }
